test: Move svn_repo fixture to general fixtures
